Agentic-Ai/task3/requirements.txt
langchain>=1.2.3
langchain-community>=0.4.1
langchain-groq>=1.1.1
langgraph>=0.1.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
